requirements.txt
asgiref==3.2.10
dataclasses==0.7
Django==3.0.5
django-appconf==1.0.4
django-cors-headers==3.4.0
django-model-utils==4.0.0
django-sendsms==0.3.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djongo==1.3.3
dnspython==2.0.0
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
